requirements.txt
pandas>=2.0
scikit-learn>=1.3
PyYAML>=6.0
matplotlib>=3.7
tqdm>=4.66
Pillow>=10.0
